replace OrderedListSpan on common class for ordered and unordered list
