+++ app/src/main/java/com/notesandroid/codex/notesandroid/note/structure/blocks/ListBlock.kt
@@ -12,7 +12,7 @@ import com.google.gson.JsonElement
 import com.notesandroid.codex.notesandroid.R
 import com.notesandroid.codex.notesandroid.Utilities.TextFormatter
 import com.notesandroid.codex.notesandroid.spans.HeightSpan
-import com.notesandroid.codex.notesandroid.spans.OrderedListSpan
+import com.notesandroid.codex.notesandroid.spans.ListSpan
 
 
 /**
@@ -23,7 +23,7 @@ import com.notesandroid.codex.notesandroid.spans.OrderedListSpan
 class ListBlock(
     context: Context,
     val data:JsonElement
-) : NoteBlock(context){
+) : NoteBlock(context) {
     override fun getView(): View = getListView()
 
 
@@ -43,15 +43,17 @@ class ListBlock(
         textView.layoutParams = param
         textView.typeface = getSameFont(R.font.roboto_regular)
         textView.setTextSize(TypedValue.COMPLEX_UNIT_SP, 14F)
-        if(data.asJsonObject["style"].asString == "ordered") {
-            val items = data.asJsonObject["items"].asJsonArray
-            for (i in 1..items.size()) {
-                val oldLength = builder.length
-                //builder.append(TextFormatter.init(context).parse(items[i - 1].asString + "\n"))
-                builder.append(TextFormatter.init(context).parse(items[i - 1].asString).append("\n"))
-                builder.setSpan(HeightSpan(25), oldLength, builder.length, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE)
-                builder.setSpan(OrderedListSpan(40, 40, i), oldLength, builder.length, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE)
-            }
+
+        val items = data.asJsonObject["items"].asJsonArray
+        for (i in 1..items.size()) {
+            val oldLength = builder.length
+            //builder.append(TextFormatter.init(context).parse(items[i - 1].asString + "\n"))
+            builder.append(TextFormatter.init(context).parse(items[i - 1].asString).append("\n"))
+            builder.setSpan(HeightSpan(25), oldLength, builder.length, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE)
+            if (data.asJsonObject["style"].asString == "ordered") {
+                builder.setSpan(ListSpan(dpToPx(20).toInt(), dpToPx(20).toInt(), i), oldLength, builder.length, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE)
+            } else
+                builder.setSpan(ListSpan(dpToPx(20).toInt(), dpToPx(20).toInt()), oldLength, builder.length, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE)
         }
 
         textView.text = builder
